Enhancement: Split test suites into unit and auto-review
