+++ .gitignore
@@ -1,3 +1,4 @@
+/.build/
 /vendor/
 /.php_cs.cache
 /infection-log.txt
