+++ .travis.yml
@@ -6,6 +6,7 @@ env:
 
 cache:
   directories:
+    - $HOME/.build/phpunit
     - $HOME/.composer/cache
     - $HOME/.php-cs-fixer
 
@@ -55,6 +56,7 @@ jobs:
         - if [[ "$WITH_HIGHEST" == "true" ]]; then composer update; fi
 
       script:
+        - vendor/bin/phpunit --configuration=test/AutoReview/phpunit.xml
         - if [[ "$WITH_COVERAGE" == "true" ]]; then xdebug-enable; fi
         - if [[ "$WITH_COVERAGE" == "true" ]]; then vendor/bin/phpunit --configuration=test/Unit/phpunit.xml --coverage-clover=build/logs/clover.xml; else vendor/bin/phpunit --configuration=test/Unit/phpunit.xml; fi
         - if [[ "$WITH_COVERAGE" == "true" ]]; then xdebug-disable; fi
