+++ Makefile
@@ -12,6 +12,7 @@ infection: vendor
 	vendor/bin/infection --min-covered-msi=80 --min-msi=80
 
 test: vendor
+	vendor/bin/phpunit --configuration=test/AutoReview/phpunit.xml
 	vendor/bin/phpunit --configuration=test/Unit/phpunit.xml
 
 vendor: composer.json composer.lock
