+++ test/AutoReview/TestCodeTest.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2017 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @link https://github.com/localheinz/factory-girl-definition
+ */
+
+namespace Localheinz\FactoryGirl\Definition\Test\AutoReview;
+
+use Localheinz\FactoryGirl\Definition\Test\Fixture;
+use Localheinz\Test\Util\Helper;
+use PHPUnit\Framework;
+
+/**
+ * @internal
+ */
+final class TestCodeTest extends Framework\TestCase
+{
+    use Helper;
+
+    public function testTestClassesAreAbstractOrFinal(): void
+    {
+        $this->assertClassesAreAbstractOrFinal(
+            __DIR__ . '/..',
+            [
+                Fixture\Definition\CanNotBeAutoloaded\MaybeUserDefinition::class,
+            ]
+        );
+    }
+}
